Create Thermometer simulation 3
For embedded systems, memory usage is a very important consideration. 
In this simulation we dynamically allocate the exact memory that shall be used by the simulation across all of its 24 readings. 24 Readings to simulate one reading at each hour of the day.
In advanced embedded systems, failure to allocate memory or improper memory allocation could lead to system crashes.